refactor: SimplifyModifier
